updated GM scripting
integrated Ilya's jumping and dashing into PlayerController
added OnPlayerDeath() func to GM
