+++ Assets/Scripts/PlayerController.cs.meta
@@ -1,7 +1,5 @@
 fileFormatVersion: 2
-guid: 83e29ce481966cc4d95acea2c0fb2c8c
-timeCreated: 1538155038
-licenseType: Free
+guid: 680548fc082e3a447ac5b4313ccc9cc4
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
